requirements.txt
mysql-connector-python 
pandas
stdiomask
